来自测试的debug
1-优化了Menu页面用户名超出显示框的情况；
2-重新排版了“历史成绩”页面
3-提交练习将隐藏提交按钮，避免重复提交
4-引入了“Master”的关于正确率的bug

+++ Entity/JudgeInfo.cs
@@ -21,8 +21,8 @@ namespace CalculateOnlineJudge.Entity
 
         public int QuestionNum { get => questionNum; set => questionNum = value; }
         public int QuestionErrorNum { get => questionErrorNum; set => questionErrorNum = value; }
-        public float CorrectRate { get => questionNum == 0 ? 0f : 1f - questionErrorNum / questionNum; }
-        public float ErrorRate { get => questionNum == 0 ? 0f : questionErrorNum / questionNum; }
+        public float CorrectRate { get => questionNum == 0 ? 0f : 1f - (float)questionErrorNum / (float)questionNum; }
+        public float ErrorRate { get => questionNum == 0 ? 0f : (float)questionErrorNum / (float)questionNum; }
         public DateTime? LastestCompleteTime { get => lastestCompleteTime; set => lastestCompleteTime = value; }
     }
 }
